# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [catalog]
      summary: List active product categories
      responses:
        "200":
          description: Active categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /products:
    get:
      operationId: getProducts
      tags: [catalog]
      summary: List active products
      responses:
        "200":
          description: Active products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/categories:
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin]
      summary: Rename a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Category not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: Duplicate category, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Remove a category without deleting products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCategoryInput"
      responses:
        "204": { description: Category removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Category not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: Replacement required or duplicate category, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /admin/products:
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/products/import:
    post:
      operationId: importProducts
      tags: [admin]
      summary: Import new products from an Excel workbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWorkbook"
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/products/{id}/stock:
    patch:
      operationId: updateProductStock
      tags: [admin]
      summary: Set product stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, active, createdAt, updatedAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      required: [name]
    DeleteCategoryInput:
      type: object
      properties:
        replacementCategoryId:
          type: string
          format: uuid
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        price:
          type: integer
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
        imageUrl:
          type: ["string", "null"]
          format: uri-reference
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, category, price, stockQuantity, imageUrl, active, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        categoryId:
          type: string
          format: uuid
        price:
          type: integer
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
        imagePath:
          type: ["string", "null"]
          maxLength: 500
      required: [name, categoryId, price, stockQuantity, imagePath]
    ProductWorkbook:
      type: object
      properties:
        fileName:
          type: string
          pattern: '\.xlsx$'
        base64:
          type: string
          minLength: 1
          maxLength: 4200000
      required: [fileName, base64]
    ProductImportResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          type: integer
        categoriesCreated:
          type: integer
      required: [created, skipped, categoriesCreated]
...
    StockUpdate:
      type: object
      properties:
        stockQuantity:
          type: integer
          minimum: 0
      required: [stockQuantity]
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Administrator role required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"