# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  categoryId: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stockQuantity: number;
  /**
     * @maxLength 500
     * @nullable
     */
  imagePath: string | null;
}