# lib/api-zod/src/generated/types/productWorkbook.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductWorkbook {
  /** @pattern \.xlsx$ */
  fileName: string;
  /**
     * @minLength 1
     * @maxLength 4200000
     */
  base64: string;
}